[BE] Feat: category, work order change API

+++ backend/src/main/java/com/ssafy/calmwave/controller/CategoryController.java
@@ -30,6 +30,7 @@ public class CategoryController {
     private final WorkCategoryRepository workCategoryRepository;
 
     /**
+     * 카테고리 생성
      * @param token
      * @param categoryRequestDto
      * @return
@@ -38,88 +39,90 @@ public class CategoryController {
     @ApiOperation(value = "카테고리 추가", notes = "result:ok, cateName,cateColor,cateIcon만 보내면 됨")
     public ResponseEntity<?> create(@RequestHeader(value = "AccessToken") String token, @RequestBody WorkCategoryDto categoryRequestDto) {
         Map<String, Object> resultMap = new HashMap<>();
-        HttpStatus status;
         User user = jwtUtil.getUser(token);
         WorkCategory byName = categoryService.findByName(categoryRequestDto.getCateName(), user);
-
-        if (byName == null) { //처음 만드는 카테고리일때
-            WorkCategory cate = WorkCategory.builder()
-                    .cateName(categoryRequestDto.getCateName())
-                    .cateColor(categoryRequestDto.getCateColor())
-                    .cateIcon(categoryRequestDto.getCateIcon())
-                    .workCategoryStatus(WorkCategoryStatus.VALID)
-                    .user(user)
-                    .build();
-            WorkCategory save = categoryService.save(cate);
-            resultMap.put("result", "ok");
-            status = HttpStatus.ACCEPTED;
-        } else if (byName != null && byName.getStatus() == WorkCategoryStatus.DELETED) {//예전에 만들었던 카테고리일때
-            byName.setStatus(WorkCategoryStatus.VALID);
-            categoryService.save(byName);
-            resultMap.put("result", "ok");
-            status = HttpStatus.ACCEPTED;
-        } else {
-            resultMap.put("result", "이미 사용중인 카테고리입니다!");
-            status = HttpStatus.INTERNAL_SERVER_ERROR;
-        }
+        HttpStatus status = categoryService.create(byName, user, categoryRequestDto);
         return new ResponseEntity<Map<String, Object>>(resultMap, status);
     }
 
     /**
+     * User별 Category를 order순으로 조회
      * @param token
      * @return List<WorkCategoryDto>
      */
     @GetMapping("list")
     @ApiOperation(value = "유저별 카테고리 리스트 order순 조회", notes = "", response = WorkCategoryDto.class)
     public ResponseEntity<?> getCategoryList(@RequestHeader(value = "AccessToken") String token) {
-        Map<String, Object> resultMap = new HashMap<>();
-        HttpStatus status;
         User user = jwtUtil.getUser(token);
         List<WorkCategoryDto> workCategories = categoryService.findByUserAndStatus(user);
         return ResponseEntity.ok().body(workCategories);
     }
 
     /**
+     * 카테고리 삭제
      * @param cateId
-     * @return ok
+     * @return "ok"
      */
     @GetMapping("delete/{cateid}")
     @ApiOperation(value = "카테고리 삭제", notes = "result:ok")
-    public ResponseEntity<?> deleteCategory(@PathVariable("cateid") Long cateId) {
+    public ResponseEntity<?> deleteCategory(@PathVariable("cateid") Long cateId, @RequestHeader(value = "AccessToken") String token) {
         Map<String, Object> resultMap = new HashMap<>();
         HttpStatus status;
-        categoryService.deleteById(cateId);
+        Optional<WorkCategory> byId = categoryService.findById(cateId);
+        //카테고리의 주인이 맞는지 확인
+        if (byId.isPresent() && categoryService.checkValid(byId,token)) {
+            categoryService.deleteById(cateId);
+            resultMap.put("result", "ok");
+            status = HttpStatus.ACCEPTED;
+        } else {
+            resultMap.put("result", "해당 카테고리를 삭제할 권한이 없습니다.");
+            status = HttpStatus.INTERNAL_SERVER_ERROR;
+        }
+        return new ResponseEntity<Map<String, Object>>(resultMap, status);
+    }
+
+    /**
+     * category 순서 변경
+     * @param cateIds
+     * @return "ok"
+     */
+    @PostMapping("/order")
+    public ResponseEntity<?> updateComponentOrder(@RequestBody List<Long> cateIds) {
+        List<WorkCategory> workCategories = workCategoryRepository.findAllById(cateIds);
+        for (int i = 0; i < workCategories.size(); i++) {
+            workCategories.get(i).setCateOrder(i);
+        }
+        workCategoryRepository.saveAll(workCategories);
+        Map<String, Object> resultMap = new HashMap<>();
         resultMap.put("result", "ok");
-        status = HttpStatus.ACCEPTED;
+        HttpStatus status = HttpStatus.ACCEPTED;
         return new ResponseEntity<Map<String, Object>>(resultMap, status);
     }
 
     /**
      * 카테고리 수정
      * @param workCategoryDto
-     * @return
+     * @return "ok"
      */
     @PostMapping("update/{cateid}")
     @ApiOperation(value = "카테고리 수정", notes = "result:ok")
-    public ResponseEntity<?> join(@RequestBody WorkCategoryDto workCategoryDto) {
+    public ResponseEntity<?> join(@RequestHeader(value = "AccessToken") String token, @RequestBody WorkCategoryDto workCategoryDto) {
         Map<String, Object> resultMap = new HashMap<>();
         HttpStatus status;
         Optional<WorkCategory> byId = categoryService.findById(workCategoryDto.getCateId());
-        WorkCategory workCategory = byId.get();
-        workCategory.setCateColor(workCategoryDto.getCateColor());
-        workCategory.setCateIcon(workCategoryDto.getCateIcon());
-        workCategory.setCateName(workCategoryDto.getCateName());
-        WorkCategory save = categoryService.save(workCategory);
-        if (save != null) {
-            //데이터 요청 성공
+        //카테고리의 주인이 맞는지 확인
+        if (byId.isPresent() && categoryService.checkValid(byId,token)) {
+            WorkCategory workCategory = byId.get();
+            workCategory.setCateColor(workCategoryDto.getCateColor());
+            workCategory.setCateIcon(workCategoryDto.getCateIcon());
+            workCategory.setCateName(workCategoryDto.getCateName());
+            categoryService.save(workCategory);
             resultMap.put("result", "ok");
             status = HttpStatus.ACCEPTED;
         } else {
-            //요청 실패
-            resultMap.put("result", "err");
+            resultMap.put("result", "해당 카테고리를 수정할 권한이 없습니다.");
             status = HttpStatus.INTERNAL_SERVER_ERROR;
         }
         return new ResponseEntity<Map<String, Object>>(resultMap, status);
     }
-
 }
